added try-catch block to ExchangeManager

+++ src/dao/ExchangeManagerDAO.js
@@ -14,7 +14,7 @@ export default class ExchangeManagerDAO extends AbstractContractDAO {
     super(
       ExchangeManagerABI,
       at,
-      MultiEventsHistoryABI
+      MultiEventsHistoryABI,
     )
   }
 
@@ -33,7 +33,7 @@ export default class ExchangeManagerDAO extends AbstractContractDAO {
         exchange.authorizedManager(),
         exchange.isActive(),
       ],
-      exchange
+      exchange,
     )
     return tx.tx
   }
@@ -47,7 +47,11 @@ export default class ExchangeManagerDAO extends AbstractContractDAO {
     let result = {}
     try {
       const assetSymbols = await exchangeProvider.getAssetSymbols()
-      assetSymbols.map((exchange) => result[web3Converter.bytesToString(exchange.symbol)] = true)
+      assetSymbols.map((exchange) => {
+        if (exchange.symbol) {
+          result[web3Converter.bytesToString(exchange.symbol)] = true
+        }
+      })
     } catch (e) {
       throw new Error(`Middleware disconnected`)
     }
@@ -68,33 +72,42 @@ export default class ExchangeManagerDAO extends AbstractContractDAO {
   async getExchangeData (exchangesAddresses: Array<string>, tokens: TokensCollection) {
     let exchangesCollection = new ExchangesCollection()
 
-    const [symbols, buyPrices, buyDecimals, sellPrices, sellDecimals, assetBalances, ethBalances] = await this._call('getExchangeData', [exchangesAddresses])
+    if (!exchangesAddresses.length) {
+      return exchangesCollection
+    }
 
-    exchangesAddresses.forEach((address, i) => {
-      const symbol = this._c.bytesToString(symbols[i])
-      const buyPrice = new BigNumber(buyPrices[i])
-      const sellPrice = new BigNumber(sellPrices[i])
-      const assetBalance = assetBalances[i]
-      const ethBalance = ethBalances[i]
-      const token = tokens.getBySymbol(symbol)
+    try {
+      const [symbols, buyPrices, buyDecimals, sellPrices, sellDecimals, assetBalances, ethBalances] = await this._call('getExchangeData', [exchangesAddresses])
 
-      try {
-        exchangeService.subscribeToExchange(address)
-        exchangeService.subscribeToToken(token, address)
-      } catch (e) {
-        // eslint-disable-next-line
-        console.error('watch error', e.message)
-      }
+      exchangesAddresses.forEach((address, i) => {
+        const symbol = this._c.bytesToString(symbols[i])
+        const buyPrice = new BigNumber(buyPrices[i])
+        const sellPrice = new BigNumber(sellPrices[i])
+        const assetBalance = assetBalances[i]
+        const ethBalance = ethBalances[i]
+        const token = tokens.getBySymbol(symbol)
 
-      exchangesCollection = exchangesCollection.add(new ExchangeOrderModel({
-        address: address,
-        symbol,
-        buyPrice: this._c.fromWei(buyPrice).mul(Math.pow(10, token.decimals())),
-        sellPrice: this._c.fromWei(sellPrice).mul(Math.pow(10, token.decimals())),
-        assetBalance: token.dao().removeDecimals(assetBalance),
-        ethBalance: this._c.fromWei(ethBalance),
-      }))
-    })
+        try {
+          exchangeService.subscribeToExchange(address)
+          exchangeService.subscribeToToken(token, address)
+        } catch (e) {
+          // eslint-disable-next-line
+          console.error('watch error', e.message)
+        }
+
+        exchangesCollection = exchangesCollection.add(new ExchangeOrderModel({
+          address: address,
+          symbol,
+          buyPrice: this._c.fromWei(buyPrice).mul(Math.pow(10, token.decimals())),
+          sellPrice: this._c.fromWei(sellPrice).mul(Math.pow(10, token.decimals())),
+          assetBalance: token.dao().removeDecimals(assetBalance),
+          ethBalance: this._c.fromWei(ethBalance),
+        }))
+      })
+    } catch (e) {
+      // eslint-disable-next-line
+      console.log(e.message)
+    }
     return exchangesCollection
   }
 
@@ -104,7 +117,7 @@ export default class ExchangeManagerDAO extends AbstractContractDAO {
   }
 
   watchExchangeCreated (account, callback) {
-    this._watch('ExchangeCreated', (tx)=>{
+    this._watch('ExchangeCreated', (tx) => {
       tx.args.symbol = this._c.bytesToString(tx.args.symbol)
       callback(tx)
     })
